fix: rename bits-of-advice to bitsofa
