+++ pages/index.tsx
@@ -194,6 +194,7 @@ export default function Home() {
             title: 'Others.',
             description: 'Everything else.',
             items: [
+                { name: 'BITS of Advice', path: '/bitsofa' },
                 { name: 'Rants', path: '/rants' },
                 { name: 'Professor Chambers', path: '/chambers' },
                 {
@@ -201,7 +202,6 @@ export default function Home() {
                     onClick: () => setShowLinksModal(true),
                 },
                 { name: 'FAQs', path: '/faqs' },
-                { name: 'BITS of Advice', path: '/bits-of-advice' },
             ],
         },
     ]
